feat: AttributeList refactor
